+++ volterra-py/src/lib.rs
@@ -21,7 +21,7 @@ mod braid;
 
 use volterra_core::ActiveNematicParams;
 use volterra_core::{QField2D, ScalarField2D, VelocityField2D};
-use volterra_solver::{
+use volterra_fd::{
     BechStats, DefectInfo, SnapStats,
     ch_step_etd,
     k0_convolution,